Fix: Change bf16 to fp16 for T4 GPU compatibility
CRITICAL BUG: ValueError: Your setup doesn't support bf16/gpu
ERROR: Training fails to start on T4 GPU

ROOT CAUSE:
- TrainingArguments configured with bf16=True
- T4 GPUs do NOT support BF16 (bfloat16) precision
- T4 only supports FP16 (float16) and FP32 (float32)
- BF16 requires Ampere architecture or newer (A100, H100)
- T4 is Turing architecture (older generation)

SOLUTION:
Changed line 260 in update_model.py:
  bf16=True    fp16=True

GPU PRECISION SUPPORT:
- T4 (Turing):     FP32,  FP16,  BF16
- A100 (Ampere):   FP32,  FP16,  BF16
- H100 (Hopper):   FP32,  FP16,  BF16,  FP8

IMPACT:
- Training now works on T4 with FP16
- Maintains memory efficiency (16-bit vs 32-bit)
- Slightly different precision than BF16 but suitable for LoRA training
- No performance degradation for ARC task

TESTED WITH:
- Google Colab T4 GPU (free tier)
- Mistral-7B-Instruct-v0.2 with 8-bit base + FP16 LoRA

+++ few-shot/arclib/update_model.py
@@ -257,7 +257,7 @@ class TTT:
             logging_steps=1,
             save_strategy="no",
             report_to="none",
-            bf16=True,  # Use bfloat16 precision
+            fp16=True,  # Use float16 precision (T4 doesn't support bf16)
             remove_unused_columns=False,
             optim="adamw_torch",
             warmup_steps=11,
